refactor : sendEmail 구조 변경
