+++ backend/src/main/java/com/ssafy/home/global/auth/constant/EmailConstraints.java
@@ -1,0 +1,10 @@
+package com.ssafy.home.global.auth.constant;
+
+public final class EmailConstraints {
+    // Util로 빼기 constant 로 넘기기 (상수
+    public static final String MAIL_TITLE = "where-is-my-home의 비밀번호 변경 이메일 입니다.";
+
+    public static final char[] charSet = { '0', '1', '2', '3', '4', '5', '6', '7', '8', '9', 'A', 'B', 'C', 'D', 'E', 'F',
+            'G', 'H', 'I', 'J', 'K', 'L', 'M', 'N', 'O', 'P', 'Q', 'R', 'S', 'T', 'U', 'V', 'W', 'X', 'Y', 'Z' };
+
+}
